requirements.txt
requests
python-dotenv
openpyxl
Pillow 